requirements.txt
Pillow==4.1.0
requests==2.13.0
imageio==2.1.2
Flask==0.12.1
